Integer som id och IDENTITY som strategy

+++ src/main/java/com/edufy/edufy/controllers/MusicController.java
@@ -32,7 +32,7 @@ public class MusicController {
     }
     // GET MUSIC BY ID todo: ResponseEntity <----------------------
     @GetMapping("/musicbyid/{id}")
-    public Optional<Music> musicById(@PathVariable("id") long id){
+    public Optional<Music> musicById(@PathVariable("id") int id){
         return musicService.getMusicById(id);
     }
 
@@ -47,7 +47,7 @@ public class MusicController {
     // DELETE MUSIC
     // returnerar en sträng som bekräftelse todo: kan ändras <-----------
     @DeleteMapping("/deletemusic/{id}")
-    public String deleteMusic(@PathVariable("id") long id){
+    public String deleteMusic(@PathVariable("id") int id){
         return musicService.deleteMusic(id);
     }
 }
